requirements.txt
# Core dependencies
streamlit==1.29.0
streamlit-oauth==0.1.0

# LLM and AI
google-generativeai==0.3.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
supabase==2.0.3

# Authentication
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Web scraping and API calls
requests==2.31.0
beautifulsoup4==4.12.2
httpx==0.25.2

# Embeddings and vector search
sentence-transformers==2.2.2
chromadb==0.4.18

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
pandas==2.1.4
plotly==5.18.0

# Rate limiting
slowapi==0.1.9

# Logging
loguru==0.7.2
